changing int() to float() to accomodate more values

+++ 2. Variables and interactive programs/2.py
@@ -1,3 +1,3 @@
-radius = int(input("Enter the radius of a circle: "))
+radius = float(input("Enter the radius of a circle: "))
 area = 3.14 * radius**2
 print("The area of the circle is: " + str(area))
